Moving armi/scripts/migration to armi/migration
