+++ armi/migration/tests/test_m0_1_6.py
@@ -15,7 +15,7 @@
 import io
 import unittest
 
-from armi.scripts.migration.m0_1_6 import ConvertAlphanumLocationSettingsToNum
+from armi.migration.m0_1_6 import ConvertAlphanumLocationSettingsToNum
 from armi.settings import caseSettings
 from armi.settings.settingsIO import SettingsWriter, SettingsReader
